refactor: export query keys from own module, export assistant hooks
