+++ packages/data-provider/src/index.ts
@@ -1,6 +1,8 @@
 export * from './data-service';
 export { default as request } from './request';
 export * from './types';
+export * from './assistants';
+export * from './query-keys';
 export * from './react-query-service';
 export * from './headers-helpers';
 export * from './sse';
