+++ packages/data-provider/src/query-keys.ts
@@ -1,0 +1,18 @@
+export enum QueryKeys {
+  messages = 'messages',
+  allConversations = 'allConversations',
+  conversation = 'conversation',
+  searchEnabled = 'searchEnabled',
+  user = 'user',
+  name = 'name', // user key name
+  models = 'models',
+  balance = 'balance',
+  endpoints = 'endpoints',
+  presets = 'presets',
+  searchResults = 'searchResults',
+  tokenCount = 'tokenCount',
+  availablePlugins = 'availablePlugins',
+  startupConfig = 'startupConfig',
+  assistants = 'assistants',
+  assistant = 'assistant',
+}
