+++ packages/data-provider/src/react-query-service.ts
@@ -9,23 +9,7 @@ import {
 import * as t from './types';
 import * as s from './schemas';
 import * as dataService from './data-service';
-
-export enum QueryKeys {
-  messages = 'messages',
-  allConversations = 'allConversations',
-  conversation = 'conversation',
-  searchEnabled = 'searchEnabled',
-  user = 'user',
-  name = 'name', // user key name
-  models = 'models',
-  balance = 'balance',
-  endpoints = 'endpoints',
-  presets = 'presets',
-  searchResults = 'searchResults',
-  tokenCount = 'tokenCount',
-  availablePlugins = 'availablePlugins',
-  startupConfig = 'startupConfig',
-}
+import { QueryKeys } from './query-keys';
 
 export const useAbortRequestWithMessage = (): UseMutationResult<
   void,
